Added option to use whitelist for non tile entity blocks to pick them.
This option will allow players to pick up any normal, non tile entity block that are in the whitelist. This is true even if tile entities are using blacklist and if pickup all blocks is enabled.

This was made to allow for picking some blocks that can be carried without having to rely on whitelist for all tile entities, as this might be time consuming task that can be avoided with this change.